Restructureing all the code
