+++ generic/piranha-tree-sitter/src/config.rs
@@ -16,501 +16,51 @@ Copyright (c) 2022 Uber Technologies, Inc.
 //! This module defines all basic building block `structs` used by Piranha.
 
 use crate::{
-  tree_sitter::TreeSitterHelpers,
-  utilities::{read_toml, MapOfVec},
+  models::{
+    outgoing_edges::{OutgoingEdges, Edges},
+    rule::{Rule, Rules},
+    scopes::{ScopeConfig, ScopeGenerator},
+  },
+  piranha::piranha_arguments::PiranhaArguments,
+  utilities::read_toml,
 };
 
-use colored::Colorize;
-use log::info;
-use serde_derive::Deserialize;
 use std::{
-  collections::HashMap,
-  hash::Hash,
+
   path::{Path, PathBuf},
 };
-use tree_sitter::{Language, Query};
 
-use self::command_line_arguments::PiranhaArguments;
+use clap::Parser;
 
-static FEATURE_FLAG_API_GROUP: &str = "Feature-flag API cleanup";
-pub static GLOBAL: &str = "Global";
-pub static METHOD: &str = "Method";
-pub static CLASS: &str = "Class";
-pub static PARENT: &str = "Parent";
+/// Used for parsing command-line arguments passed to Piranha .
+#[derive(Clone, Parser, Debug)]
+#[clap(author, version, about, long_about = None)]
 
-pub mod command_line_arguments {
-  //! This module contains `structs` and implementations for parsing and managing command line arguments passed to Piranha.
-  use clap::Parser;
-  use colored::Colorize;
-  use log::info;
-  use serde_derive::Deserialize;
-  use std::{collections::HashMap, path::PathBuf};
-  use tree_sitter::Language;
-
-  use crate::{tree_sitter::TreeSitterHelpers, utilities::read_toml};
-
-  /// Used for parsing command-line arguments passed to Piranha .
-  #[derive(Clone, Parser, Debug)]
-  #[clap(author, version, about, long_about = None)]
-
-  pub struct CommandLineArguments {
-    /// Path to source code folder.
-    #[clap(short = 'c', long)]
-    pub path_to_codebase: String,
-    /// Folder containing the required configuration files
-    #[clap(short = 'f', long)]
-    pub path_to_feature_flag_rules: String,
-    /// Path to the file containing arguments for Piranha
-    #[clap(short = 'p', long)]
-    pub path_to_piranha_arguments: String,
-  }
-
-  /// Captures the Piranha arguments by from the file at `path_to_feature_flag_rules`.
-  #[derive(Deserialize, Debug, Clone, Hash, PartialEq, Eq, Default)]
-  struct PiranhaArgsFromConfig {
-    language: Vec<String>,
-    substitutions: Vec<Vec<String>>,
-  }
-
-  #[derive(Clone)]
-  /// Captures the processed Piranha arguments (PiranhaArgsFromConfig) that are parsed from `path_to_feature_flag_rules`.
-  pub struct PiranhaArguments {
-    /// Path to source code folder.
-    path_to_code_base: String,
-    // Input arguments provided to Piranha, mapped to tag names -
-    // @stale_flag_name, @namespace, @treated, @treated_complement
-    // These substitutions instantiate the initial set of feature flag rules.
-    input_substitutions: HashMap<String, String>,
-    /// Folder containing the API specific rules
-    path_to_configurations: String,
-    /// Tree-sitter language model
-    language: Language,
-    // The language name is file the extension used for files in particular language.
-    language_name: String,
-  }
-
-  impl PiranhaArguments {
-    pub fn new(args: CommandLineArguments) -> Self {
-      let path_to_piranha_argument_file = PathBuf::from(args.path_to_piranha_arguments.as_str());
-
-      let piranha_args_from_config: PiranhaArgsFromConfig =
-        read_toml(&path_to_piranha_argument_file, false);
-
-      let input_substitutions = piranha_args_from_config
-        .substitutions
-        .iter()
-        .map(|x| (String::from(&x[0]), String::from(&x[1])))
-        .collect();
-
-      #[rustfmt::skip]
-      info!("{}",  format!("Piranha arguments are :\n {:?}", input_substitutions).purple());
-
-      Self {
-        path_to_code_base: args.path_to_codebase.to_string(),
-        input_substitutions,
-        path_to_configurations: args.path_to_feature_flag_rules.to_string(),
-        language_name: String::from(&piranha_args_from_config.language[0]),
-        language: piranha_args_from_config.language[0].get_language(),
-      }
-    }
-
-    pub fn path_to_code_base(&self) -> &str {
-      self.path_to_code_base.as_ref()
-    }
-
-    pub fn input_substitutions(&self) -> &HashMap<String, String> {
-      &self.input_substitutions
-    }
-
-    pub fn path_to_configurations(&self) -> &str {
-      self.path_to_configurations.as_ref()
-    }
-
-    pub fn language(&self) -> Language {
-      self.language
-    }
-
-    pub fn language_name(&self) -> &str {
-      self.language_name.as_ref()
-    }
-  }
+pub struct CommandLineArguments {
+  /// Path to source code folder.
+  #[clap(short = 'c', long)]
+  pub path_to_codebase: String,
+  /// Folder containing the required configuration files
+  #[clap(short = 'f', long)]
+  pub path_to_feature_flag_rules: String,
+  /// Path to the file containing arguments for Piranha
+  #[clap(short = 'p', long)]
+  pub path_to_piranha_arguments: String,
 }
 
-#[derive(Deserialize, Debug, Clone, Hash, PartialEq, Eq, Default)]
-pub struct Rule {
-  /// Name of the rule. (It is unique)
-  name: String,
-  /// Tree-sitter query as string
-  query: String,
-  /// The tag corresponding to the node to be replaced
-  replace_node: String,
-  /// Replacement pattern
-  replace: String,
-  /// Group(s) to which the rule belongs
-  groups: Option<Vec<String>>,
-  /// Holes that need to be filled, in order to instantiate a rule
-  holes: Option<Vec<String>>,
-  /// Additional constraints for matching the rule
-  constraints: Option<Vec<Constraint>>,
-  /// Heuristics for identifying potential files containing occurrence of the rule.
-  grep_heuristics: Option<Vec<String>>,
-}
-
-#[derive(Deserialize, Debug, Clone, Hash, PartialEq, Eq)]
-pub struct Constraint {
-  /// Scope in which the constraint query has to be applied
-  matcher: String,
-  /// The Tree-sitter queries that need to be applied in the `matcher` scope
-  queries: Vec<String>,
-}
-
-impl Constraint {
-  pub fn queries(&self) -> &[String] {
-    &self.queries
-  }
-
-  pub fn matcher(&self) -> String {
-    String::from(&self.matcher)
-  }
-}
-
-impl Rule {
-  /// Create a new query from `self` with the input `query` and `replace`
-  pub fn update(
-    &self, query: String, replace: String, substitutions: &HashMap<String, String>,
-  ) -> Result<Rule, String> {
-    if substitutions.len() != self.holes().len() {
-      #[rustfmt::skip]
-      return Err(format!("Could not instantiate a rule - {:?}. Some Holes {:?} not found in table {:?}", self, self.holes(), substitutions));
-    } else {
-      let mut updated_rule = self.clone();
-      if !updated_rule.holes().is_empty() {
-        updated_rule.set_query(query.substitute_tags(substitutions));
-        updated_rule.set_replace(replace.substitute_tags(substitutions));
-      }
-      Ok(updated_rule)
-    }
-  }
-
-  fn is_feature_flag_cleanup(&self) -> bool {
-    self.groups().iter().any(|t| t.eq(FEATURE_FLAG_API_GROUP))
-  }
-
-  // Dummy rules are helper rules that make it easier to define the rule graph.
-  fn is_dummy_rule(&self) -> bool {
-    return self.query.is_empty() && self.replace.is_empty();
-  }
-
-  /// Instantiate `self` with substitutions or panic.
-  pub fn instantiate(&self, substitutions: &HashMap<String, String>) -> Rule {
-    if let Ok(r) = self.try_instantiate(substitutions) {
-      return r;
-    }
-    #[rustfmt::skip]
-    panic!("{}", format!("Could not instantiate the rule {:?} with substitutions {:?}", self, substitutions).red());
-  }
-
-  /// Groups the rules based on the field `rule.groups`
-  /// Note: a rule can belong to more than one group.
-  fn get_grouped_rules(rules: &Vec<Rule>) -> (HashMap<String, Rule>, HashMap<String, Vec<String>>) {
-    let mut rules_by_name = HashMap::new();
-    let mut rules_by_group = HashMap::new();
-    for rule in rules {
-      rules_by_name.insert(rule.name.to_string(), rule.clone());
-      for tag in rule.groups() {
-        rules_by_group.collect(tag.to_string(), rule.name.to_string());
-      }
-    }
-    (rules_by_name, rules_by_group)
-  }
-
-  /// Tries to instantiate the rule (`self`) based on the substitutions.
-  /// Note this could fail if the `substitutions` doesn't contain mappings for each hole.
-  fn try_instantiate(&self, substitutions: &HashMap<String, String>) -> Result<Rule, String> {
-    let relevant_substitutions = self
-      .holes()
-      .iter()
-      .filter_map(|hole| substitutions.get(hole).map(|subs| (hole, subs)))
-      .map(|(a, b)| (a.clone(), b.clone()))
-      .collect::<HashMap<String, String>>();
-    self.update(self.query(), self.replace(), &relevant_substitutions)
-  }
-
-  /// Records the string that should be grepped in order to find files that
-  /// potentially could match this global rule.
-  fn add_grep_heuristics_for_global_rules(&mut self, substitutions: &HashMap<String, String>) {
-    let mut gh = vec![];
-    for hole in self.holes() {
-      if let Some(x) = substitutions.get(&hole) {
-        gh.push(x.clone());
-      }
-    }
-    self.grep_heuristics = Some(gh.clone());
-  }
-
-  pub fn get_query(&self) -> String {
-    self.query.clone()
-  }
-
-  /// Adds the rule to a new group - "Feature-flag API cleanup"
-  fn add_to_feature_flag_api_group(&mut self) {
-    let group_name: String = FEATURE_FLAG_API_GROUP.to_string();
-    match self.groups.as_mut() {
-      None => self.groups = Some(vec![group_name]),
-      Some(_groups) => _groups.push(group_name),
-    }
-  }
-
-  pub fn replace_node(&self) -> String {
-    String::from(&self.replace_node)
-  }
-
-  fn query(&self) -> String {
-    String::from(&self.query)
-  }
-
-  pub fn replace(&self) -> String {
-    String::from(&self.replace)
-  }
-
-  pub fn constraints(&self) -> Vec<Constraint> {
-    match &self.constraints {
-      Some(cs) => cs.clone(),
-      None => vec![],
-    }
-  }
-
-  pub fn grep_heuristics(&self) -> Vec<String> {
-    match &self.grep_heuristics {
-      Some(cs) => cs.clone(),
-      None => vec![],
-    }
-  }
-
-  fn holes(&self) -> Vec<String> {
-    match &self.holes {
-      Some(cs) => cs.clone(),
-      None => vec![],
-    }
-  }
-
-  fn groups(&self) -> Vec<String> {
-    match &self.groups {
-      Some(cs) => cs.clone(),
-      None => vec![],
-    }
-  }
-
-  pub fn set_replace(&mut self, replace: String) {
-    self.replace = replace;
-  }
-
-  pub fn set_query(&mut self, query: String) {
-    self.query = query;
-  }
-}
-
-// Captures an entry from the `edges.toml` file.
-#[derive(Deserialize, Debug, Clone, Hash, PartialEq, Eq, Default)]
-struct Edge {
-  from: String,
-  to: Vec<String>,
-  scope: String, //
-}
-
-#[derive(Deserialize, Debug, Clone, Hash, PartialEq, Eq, Default)]
-// Represents the `edges.toml` file
-struct Edges {
-  edges: Vec<Edge>,
-}
-
-#[derive(Deserialize, Debug, Clone, Hash, PartialEq, Eq, Default)]
-// Represents the `rules.toml` file
-pub struct Rules {
-  rules: Vec<Rule>,
-}
-
-/// This maintains the state for Piranha.
-pub struct RuleStore {
-  // A graph that captures the flow amongst the rules
-  rule_graph: RuleGraph,
-  // Caches the compiled tree-sitter queries.
-  rule_query_cache: HashMap<String, Query>,
-  // All the input rules stored by name
-  rules_by_name: HashMap<String, Rule>,
-  // Current global rules to be applied.
-  global_rules: Vec<Rule>,
-  // Scope generators.
-  scopes: Vec<ScopeGenerator>,
-  // Command line arguments passed to piranha
-  piranha_args: PiranhaArguments,
-}
-
-impl RuleStore {
-  pub fn new(args: &PiranhaArguments) -> RuleStore {
-    let (rule_graph, rules_by_name, scopes) = read_rule_graph_from_config(&args);
-    let mut rule_store = RuleStore {
-      rule_graph,
-      rule_query_cache: HashMap::new(),
-      rules_by_name,
-      global_rules: vec![],
-      scopes,
-      piranha_args: args.clone(),
-    };
-
-    for (_, rule) in rule_store.rules_by_name.clone() {
-      if rule.is_feature_flag_cleanup() {
-        rule_store.add_to_global_rules(&rule, &args.input_substitutions());
-      }
-    }
-    return rule_store;
-  }
-
-  pub fn global_rules(&self) -> Vec<Rule> {
-    self.global_rules.clone()
-  }
-
-  pub fn language(&self) -> Language {
-    self.piranha_args.language()
-  }
-
-  pub fn language_name(&self) -> &str {
-    self.piranha_args.language_name()
-  }
-
-  pub fn input_substitutions(&self) -> HashMap<String, String> {
-    self.piranha_args.input_substitutions().clone()
-  }
-
-  /// Add a new global rule, along with grep heuristics.
-  pub fn add_to_global_rules(&mut self, rule: &Rule, tag_captures: &HashMap<String, String>) {
-    if let Ok(mut r) = rule.try_instantiate(&tag_captures) {
-      r.add_grep_heuristics_for_global_rules(&tag_captures);
-      #[rustfmt::skip]
-      info!("{}", format!("Added Global Rule : {:?} - {}", r.name, r.get_query()).bright_blue());
-      self.global_rules.push(r);
-    }
-  }
-
-  /// Get the compiled query for the `query_str` from the cache
-  /// else compile it, add it to the cache and return it.
-  pub fn get_query(&mut self, query_str: &String) -> &Query {
-    let language = self.language();
-    self
-      .rule_query_cache
-      .entry(query_str.to_string())
-      .or_insert_with(|| query_str.create_query(language))
-  }
-
-  /// Get the next rules to be applied grouped by the scope in which they should be performed.
-  pub fn get_next(
-    &self, rule: &Rule, tag_matches: &HashMap<String, String>,
-  ) -> HashMap<String, Vec<Rule>> {
-    let rule_name = &rule.name;
-    let mut next_rules: HashMap<String, Vec<Rule>> = HashMap::new();
-    // Iterate over each entry (Edge) in the adjacency list corresponding to `rule_name`
-    for (scope, to_rule) in self.rule_graph.get_neighbors(rule_name) {
-      let to_rule_name = &self.rules_by_name[&to_rule];
-      // If the to_rule_name is a dummy rule, skip it and rather return it's next rules.
-      if to_rule_name.is_dummy_rule() {
-        // Call this method recursively on the dummy node
-        for (next_next_rules_scope, next_next_rules) in self.get_next(to_rule_name, tag_matches) {
-          for next_next_rule in next_next_rules {
-            // Group the next rules based on the scope
-            next_rules.collect(
-              String::from(&next_next_rules_scope),
-              next_next_rule.instantiate(&tag_matches),
-            )
-          }
-        }
-      } else {
-        // Group the next rules based on the scope
-        next_rules.collect(String::from(&scope), to_rule_name.instantiate(&tag_matches));
-      }
-    }
-    // Add empty entry, incase no next rule was found for a particular scope
-    for scope in [PARENT, METHOD, CLASS, GLOBAL] {
-      next_rules.entry(scope.to_string()).or_default();
-    }
-    next_rules
-  }
-
-  // For the given scope level, get the ScopeQueryGenerator from the `scope_config.toml` file
-  pub fn get_scope_query_generators(&self, scope_level: &str) -> Vec<ScopeQueryGenerator> {
-    self
-      .scopes
-      .iter()
-      .find(|level| level.name.eq(scope_level))
-      .map(|scope| scope.rules.clone())
-      .unwrap_or_else(Vec::new)
-  }
-}
-
-/// Captures the relationship between the rules as a graph (adjacency list)
-pub struct RuleGraph(HashMap<String, Vec<(String, String)>>);
-
-impl RuleGraph {
-  // Constructs a graph of rules based on the input `edges` that represent the relationship between two rules or groups of rules.
-  fn new(edges: Vec<Edge>, all_rules: Vec<Rule>) -> Self {
-    let (rules_by_name, rules_by_group) = Rule::get_grouped_rules(&all_rules);
-
-    // A closure that gets the rules corresponding to the given rule name or group name.
-    let get_rules_for_tag_or_name = |val: &String| {
-      rules_by_name
-        .get(val)
-        .map(|v| vec![v.name.clone()])
-        .unwrap_or_else(|| rules_by_group[val].clone())
-    };
-
-    let mut graph = HashMap::new();
-    // Add the edge(s) to the graph. Multiple edges will be added
-    // when either edge endpoint is a group name.
-    for edge in edges {
-      for from_rule in get_rules_for_tag_or_name(&edge.from) {
-        for to_edge in &edge.to {
-          for t in get_rules_for_tag_or_name(&to_edge) {
-            // Add edge to the adjacency list
-            graph.collect(from_rule.clone(), (String::from(&edge.scope), t.clone()));
-          }
-        }
-      }
-    }
-    RuleGraph(graph)
-  }
-
-  /// Get all the outgoing edges for `rule_name`
-  fn get_neighbors(&self, rule_name: &String) -> Vec<(String, String)> {
-    self
-      .0
-      .get(rule_name)
-      .map(|x| x.clone())
-      .unwrap_or_else(|| vec![])
-  }
-}
-
-/// Read the language specific cleanup rules.
-fn get_cleanup_rules(language: &str) -> (Rules, Edges, Vec<ScopeGenerator>) {
-  let project_root = PathBuf::from(env!("CARGO_MANIFEST_DIR"));
-  let path_to_lang_config = &project_root.join(format!("src/cleanup_rules/{}", language));
-
-  match language {
-    "java" => (
-      read_toml(&path_to_lang_config.join("rules.toml"), false),
-      read_toml(&path_to_lang_config.join("edges.toml"), false),
-      read_toml::<ScopeConfig>(&path_to_lang_config.join("scope_config.toml"), false).scopes,
-    ),
-    _ => panic!(),
-  }
-}
-
-/// Reads the input configurations and creates a rule graph.
-fn read_rule_graph_from_config(
+pub(crate) fn read_config_files(
   args: &PiranhaArguments,
-) -> (RuleGraph, HashMap<String, Rule>, Vec<ScopeGenerator>) {
+) -> (Vec<Rule>, Vec<OutgoingEdges>, Vec<ScopeGenerator>) {
   let path_to_config = Path::new(args.path_to_configurations());
+  let project_root = PathBuf::from(env!("CARGO_MANIFEST_DIR"));
+  let path_to_lang_config = &project_root.join(format!("src/cleanup_rules/{}", args.language_name()));
+
 
   // Read the language specific cleanup rules and edges
-  let (language_rules, language_edges, scopes) = get_cleanup_rules(&args.language_name());
+  let language_rules : Rules = read_toml(&path_to_lang_config.join("rules.toml"), false);
+  let language_edges : Edges = read_toml(&path_to_lang_config.join("edges.toml"), false);
+  let scopes = read_toml::<ScopeConfig>(&path_to_lang_config.join("scope_config.toml"), false).scopes();
+  
 
   // Read the API specific cleanup rules and edges
   let (mut input_rules, input_edges): (Rules, Edges) = (
@@ -526,38 +76,5 @@ fn read_rule_graph_from_config(
   let all_rules = [language_rules.rules, input_rules.rules].concat();
   let all_edges = [language_edges.edges, input_edges.edges].concat();
 
-  let (rules_by_name, _) = Rule::get_grouped_rules(&all_rules);
-
-  let graph = RuleGraph::new(all_edges, all_rules);
-
-  (graph, rules_by_name, scopes)
-}
-
-// Represents the content in the `scope_config.toml` file
-#[derive(Deserialize, Debug, Clone, Hash, PartialEq, Eq, Default)]
-struct ScopeConfig {
-  scopes: Vec<ScopeGenerator>,
-}
-
-// Represents an entry in the `scope_config.toml` file
-#[derive(Deserialize, Debug, Clone, Hash, PartialEq, Eq, Default)]
-struct ScopeGenerator {
-  name: String,
-  rules: Vec<ScopeQueryGenerator>,
-}
-
-#[derive(Deserialize, Debug, Clone, Hash, PartialEq, Eq, Default)]
-pub struct ScopeQueryGenerator {
-  matcher: String, // a tree-sitter query matching some enclosing AST pattern (like method or class)
-  generator: String, // a tree-sitter query matching the exact AST node
-}
-
-impl ScopeQueryGenerator {
-  pub fn matcher(&self) -> String {
-    String::from(&self.matcher)
-  }
-
-  pub fn generator(&self) -> String {
-    String::from(&self.generator)
-  }
+  (all_rules, all_edges, scopes)
 }
