+++ generic/piranha-tree-sitter/src/models/rule_graph.rs
@@ -1,0 +1,49 @@
+use crate::{
+  models::{outgoing_edges::OutgoingEdges, rule::Rule},
+  utilities::MapOfVec,
+};
+use std::collections::HashMap;
+
+pub struct RuleGraph(HashMap<String, Vec<(String, String)>>);
+
+impl RuleGraph {
+  // Constructs a graph of rules based on the input `edges` that represent the relationship between two rules or groups of rules.
+  pub(crate) fn new(edges: &Vec<OutgoingEdges>, all_rules: &Vec<Rule>) -> Self {
+    let (rules_by_name, rules_by_group) = Rule::get_grouped_rules(&all_rules);
+
+    // A closure that gets the rules corresponding to the given rule name or group name.
+    let get_rules_for_tag_or_name = |val: &String| {
+      rules_by_name
+        .get(val)
+        .map(|v| vec![v.name()])
+        .unwrap_or_else(|| rules_by_group[val].clone())
+    };
+
+    let mut graph = HashMap::new();
+    // Add the edge(s) to the graph. Multiple edges will be added
+    // when either edge endpoint is a group name.
+    for edge in edges {
+      for from_rule in get_rules_for_tag_or_name(&edge.from_rule()) {
+        for outgoing_edge in edge.to_rules() {
+          for to_rule in get_rules_for_tag_or_name(&outgoing_edge) {
+            // Add edge to the adjacency list
+            graph.collect(
+              from_rule.clone(),
+              (String::from(edge.scope()), to_rule.clone()),
+            );
+          }
+        }
+      }
+    }
+    RuleGraph(graph)
+  }
+
+  /// Get all the outgoing edges for `rule_name`
+  pub(crate) fn get_neighbors(&self, rule_name: &String) -> Vec<(String, String)> {
+    self
+      .0
+      .get(rule_name)
+      .map(|x| x.clone())
+      .unwrap_or_else(|| vec![])
+  }
+}
