+++ generic/piranha-tree-sitter/test-resources/java/expected_control/XPFlagCleanerInterfaceMethod.java
