+++ generic/piranha-tree-sitter/test-resources/java/expected_treated/ExperimentInterfaceAnnotatedInterface.java
