+++ generic/piranha-tree-sitter/test-resources/java/expected_treated/XPFlagCleanerInterfaceMethod.java
