+++ generic/piranha-tree-sitter/test-resources/java/input/ExperimentInterfaceAnnotatedInterface.java
