+++ generic/piranha-tree-sitter/test-resources/java/input/OtherXPFlagCleanerInterfaceMethod.java
