+++ generic/piranha-tree-sitter/test-resources/java/input/XPFlagCleanerInterfaceMethod.java
