+++ generic/piranha-tree-sitter/test-resources/java/input/XPFlagCleanerPositiveCases.java
